Rename to aiContributorSummary
